Refactor to an actor model.  watchRucioForNewDatasets will be responsible for polling rucio with a given period, and launching action scripts when a matching dataset is discovered.
The polling period is a moving window which (by default) begins at
program execution, and ends with each iteration of the polling loop.

NOTE: There will be small gaps between each polling period, as there are
instructions which must execute between each iteration... but these
gaps are small.  And the expectation is that not all datasets / runs
will be processed during an experimental run.

Each actor is a script (or other executable object) provided by the
user.  If any actor cannot be discovered when the program starts, an
exception is raised and no actions will be taken.

Polling periods are specified in seconds.  [TODO: Issue warning
if polling is shorter than the site default and require a user
confirmation].