[pandaopts] Add pfnList option.
The user can specify a pfnList.  When specified, the contents of the
file which is referenced will be copied into the submission directory,
and from there to the PanDA server and worker node.  (Thus the contents
must be lightweight and cannot be, e.g., datafiles).  This is meant to
deal with the case where local files are used to spawn jobs.